@@ -1,6 +1,16 @@
-import { Component, OnInit } from '@angular/core';
-import { Items } from '../../constants/mockup-data';
-import { InteractionService } from 'src/app/services/interaction.service';
+import {
+  ProductsService
+} from './../../services/products.service';
+import {
+  Component,
+  OnInit
+} from '@angular/core';
+import {
+  Items
+} from '../../constants/mockup-data';
+import {
+  InteractionService
+} from 'src/app/services/interaction.service';
 
 @Component({
   selector: 'app-items',
@@ -10,15 +20,36 @@ import { InteractionService } from 'src/app/services/interaction.service';
 export class ItemsComponent implements OnInit {
   isSidePanelExpanded: boolean;
   searchText;
-  items: { name: string; unit: string; description: string; variety: string[]; }[];
+  items: {
+    name: string;unit: string;description: string;variety: string[];
+  } [];
+  allProducts: any;
 
-  constructor(private interaction: InteractionService) {
+  constructor(private interaction: InteractionService, private productService: ProductsService) {
     this.isSidePanelExpanded = this.interaction.getExpandedStatus();
+    this.productService.getAllProducts().subscribe((res) => {
+      this.allProducts = res;
+      console.log(res);
+    });
+  }
+
+  addProduct() {
+    this.productService.addProduct({
+      image: '@/Users/vijayanand/Downloads/cult.png',
+      name: 'sdscdf',
+      category: 'dscsdc',
+      details: 'csdcds',
+      sub_category: 'sscsdcc',
+      variant: 'scsdc',
+      brand: 'fdsfd',
+      quantity_type: 'scdsd',
+      variants: ['dscsd', 'csdcs']
+    });
   }
 
   ngOnInit() {
     this.items = Items;
-    this.interaction.expandedStatus$.subscribe( (res) => {
+    this.interaction.expandedStatus$.subscribe((res) => {
       this.isSidePanelExpanded = res;
     });
   }
