@@ -1,10 +1,11 @@
+import { RetailerService } from './../../services/retailer.service';
 import {Component, OnInit, ViewChild} from '@angular/core';
 import { InteractionService } from 'src/app/services/interaction.service';
 import {MatSort} from '@angular/material/sort';
 import {MatPaginator} from '@angular/material/paginator';
 import {MatTableDataSource} from '@angular/material/table';
 
-export interface PeriodicElement {
+export interface Retailer {
   name: string;
   email: string;
   phone: number;
@@ -13,7 +14,7 @@ export interface PeriodicElement {
   recentActivity: string;
 }
 
-const ELEMENT_DATA: PeriodicElement[] = [
+const ELEMENT_DATA: Retailer[] = [
   {name: 'Pranav', email: 'pranavpace97@gmail.com', phone: 9206877149, shop: 'Webknot', totalBusiness: 50000, recentActivity: '5 min' },
   {name: 'Pranav', email: 'pranavpace97@gmail.com', phone: 9206877149, shop: 'Webknot', totalBusiness: 50000, recentActivity: '5 min' },
   {name: 'Pranav', email: 'pranavpace97@gmail.com', phone: 9206877149, shop: 'Webknot', totalBusiness: 50000, recentActivity: '5 min' },
@@ -32,12 +33,18 @@ const ELEMENT_DATA: PeriodicElement[] = [
 export class RetailerComponent implements OnInit {
   isSidePanelExpanded: boolean;
   displayedColumns: string[] = ['name', 'email', 'phone', 'shop', 'totalBusiness', 'recentActivity'];
-  dataSource = new MatTableDataSource(ELEMENT_DATA);
+  dataSource;
   @ViewChild(MatSort, {static: true}) sort: MatSort;
   @ViewChild(MatPaginator, {static: true}) paginator: MatPaginator;
+  allRetailers: any;
 
-  constructor(private interaction: InteractionService) {
+  constructor(private interaction: InteractionService, private retailerService: RetailerService) {
     this.isSidePanelExpanded = this.interaction.getExpandedStatus();
+    this.retailerService.getAllRetailers().subscribe((res) => {
+      console.log(res);
+      this.allRetailers = res;
+    });
+    this.dataSource = new MatTableDataSource(ELEMENT_DATA);
   }
 
   ngOnInit() {
